+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 
+## Unreleased
+
+### Documentation
+
+- Plan the Drive modifiedTime preservation and repair
+
 ## v0.4.0 - 2026-08-13
 
 ### Features